device/requirements.txt
numpy==1.26.4
opencv-python==4.10.0.82
python-socketio[client]==5.11.2
tensorflow==2.16.1
